Channel part is added!
Now the channel part is added. When a user chooses a server and wants to
create a channel or enter one channel, a socket is created between two
channel handlers and they communicate with each other from requests. Now
the options of the channel must be added and after that, the main part
of each channel: Chatting!